Time: 251 ms (85.13%), Space: 198.4 MB (82.43%) - LeetHub

+++ 1973-count-nodes-equal-to-sum-of-descendants/1973-count-nodes-equal-to-sum-of-descendants.cpp
@@ -17,7 +17,7 @@ private:
         
         long left = dfs(curr -> left), right = dfs(curr -> right);
         
-        if (curr -> val == (long)left + (long)right) ++count;
+        if (curr -> val == left + right) ++count;
         
         return left + right + curr -> val;
     }
